Rename internal/errs to internal/colerrs
